Delete duplicates test https://github.com/RedRoverSchool/JenkinsQA_2024_fall/issues/1336
